Move LogWrapper to android package
