+++ audioswitch/src/main/java/com/twilio/audioswitch/WiredHeadsetReceiver.kt
@@ -4,6 +4,7 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import android.content.IntentFilter
+import com.twilio.audioswitch.android.LogWrapper
 
 private const val TAG = "WiredHeadsetReceiver"
 internal const val STATE_UNPLUGGED = 0
