+++ audioswitch/src/main/java/com/twilio/audioswitch/android/LogWrapper.kt
@@ -1,4 +1,4 @@
-package com.twilio.audioswitch
+package com.twilio.audioswitch.android
 
 import android.util.Log
